# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Databox — data workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blocks/{blockId}/todo-items:
    parameters:
      - name: blockId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listTodoItems
      tags: [todo-items]
      summary: List todo items in a block
      responses:
        "200":
          description: List of todo items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoItem"
    post:
      operationId: createTodoItem
      tags: [todo-items]
      summary: Create a todo item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoItemInput"
      responses:
        "201":
          description: Created todo item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoItem"
  /todo-items/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateTodoItem
      tags: [todo-items]
      summary: Update a todo item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoItemUpdate"
      responses:
        "200":
          description: Updated todo item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTodoItem
      tags: [todo-items]
      summary: Delete a todo item
      responses:
        "204":
          description: Deleted

  # ─── Calendar Events ──────────────────────────────────────────────────────
...
  /blocks/{blockId}/photos:
    parameters:
      - name: blockId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List photos in a block
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
    post:
      operationId: addPhoto
      tags: [photos]
      summary: Add a photo record (after upload to storage)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "201":
          description: Created photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
  /photos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updatePhoto
      tags: [photos]
      summary: Update a photo caption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoUpdate"
      responses:
        "200":
          description: Updated photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      responses:
        "204":
          description: Deleted

  # ─── Analysis ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Categories ────────────────────────────────────────────────────────
...
    TodoItem:
      type: object
      properties:
        id:
          type: integer
        blockId:
          type: integer
        text:
          type: string
        completed:
          type: boolean
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, blockId, text, completed, position, createdAt, updatedAt]
    TodoItemInput:
      type: object
      properties:
        text:
          type: string
      required: [text]
    TodoItemUpdate:
      type: object
      properties:
        text:
          type: string
        completed:
          type: boolean

    # ── Calendar Events ───────────────────────────────────────────────────
...
    Photo:
      type: object
      properties:
        id:
          type: integer
        blockId:
          type: integer
        objectPath:
          type: string
        caption:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, blockId, objectPath, createdAt, updatedAt]
    PhotoInput:
      type: object
      properties:
        objectPath:
          type: string
        caption:
          type: ["string", "null"]
      required: [objectPath]
    PhotoUpdate:
      type: object
      properties:
        caption:
          type: ["string", "null"]

    # ── Analysis ──────────────────────────────────────────────────────────
...