# lib/api-zod/src/generated/types/todoItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Databox — data workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface TodoItemUpdate {
  text?: string;
  completed?: boolean;
}